+++ Content.Server/Ninja/Systems/SpaceNinjaSystem.cs
@@ -15,7 +15,6 @@ using Content.Shared.PowerCell;
 using Content.Shared.Popups;
 using Content.Shared.Rounding;
 using System.Diagnostics.CodeAnalysis;
-using Content.Server._Starlight.Ninja;
 using Content.Shared._Starlight.Ninja;
